add new filter for adding prefixes
new 'opengraph_prefixes' filter allows additional prefix declarations to
be added easily.

Partially fixes #7.  Ideally, even though numerous prefixes may be
registered, only the ones used on a given page would be output.
Conceptually, this would be similar to the register and enqueue
distinction of scripts and styles in WordPress