populate meta 'name' attribute alongside 'property'
While a slight abuse of the original intent of this plugin, this allows
extensions to register new properties in the "twitter:" namepsace and
have them picked up as Twitter Cards
(https://dev.twitter.com/docs/cards).  Becase both the "og:*" and
"twitter:*" names are actually registered at
http://wiki.whatwg.org/wiki/MetaExtensions, this stays compliant with
the HTML5 spec.