LICENSE: update to correct license
This changes the license in the LICENSE from MIT to Apache 2.0.
The plugin was always originally Apache 2.0 licensed (as can be seen in
the very first commit: ec018f6).

I honestly don't know why I put an MIT license in the LICENSE file in
47176b1, and am a little surprised in went unnoticed for so long. Thanks
to @pfefferle for pointing it out in #35.